cleaned up campground
